wiregrid_encoder: Move wg encoder agent into package
