+++ docker-compose.yml
@@ -113,15 +113,6 @@ services:
       - "socs"
 
   # --------------------------------------------------------------------------
-  # Wiregrid encoder agent
-  # --------------------------------------------------------------------------
-  ocs-wgencoder-agent:
-    image: "ocs-wgencoder-agent"
-    build: ./agents/wiregrid_encoder/
-    depends_on:
-      - "socs"
-
-  # --------------------------------------------------------------------------
   # iBootbar agent
   # --------------------------------------------------------------------------
   ocs-ibootbar-agent:
